* benchmark/prepare_require.rb: skip file creation if it already   exist. Suggested by ko1.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@52086 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ benchmark/prepare_require.rb
@@ -1,14 +1,25 @@
 require "fileutils"
 
-basename = File.dirname($0)
-data_dir = File.join(basename, "bm_require.data")
+def prepare
+  num_files = 10000
 
-FileUtils.mkdir_p(data_dir)
+  basename = File.dirname($0)
+  data_dir = File.join(basename, "bm_require.data")
 
-1.upto(10000) do |i|
-  f = File.open("#{data_dir}/c#{i}.rb", "w")
-  f.puts <<-END
+  # skip if all of files exists
+  if File.exist?(File.join(data_dir, "c#{num_files}.rb"))
+    return
+  end
+
+  FileUtils.mkdir_p(data_dir)
+
+  1.upto(num_files) do |i|
+    f = File.open("#{data_dir}/c#{i}.rb", "w")
+    f.puts <<-END
       class C#{i}
       end
     END
+  end
 end
+
+prepare
